requirements_3.txt
aiobotocore
aiohttp
aiodns
boto3
cchardet
urllib3
opencv-python
Pillow
scikit-image
lailib
Cerberus
pyclipper
editdistance
tqdm
cython
pycocotools
schema
timm
omegaconf
pytorch-lightning
onnxruntime
numpy
msgpack
msgpack_numpy
autograd
scikit-learn
func-timeout
CairoSVG
pandas
onnx
protobuf
matplotlib
seaborn
shapely
p_tqdm
pika
aio_pika